CHANGES WERE MADE TO MakeSheet.cs MAINLY (some exceptions) --------------------------------------------------------- CLEANUP-CHANGES: Made some more minor changes, making the code smaller.  Examples of this, are that the event handlers have been a bit consolidated, due to their nature being the same.  And a few other pieces of code that were consolidated as well.  Made some changes to the EXP, though it seems it doesn't fit with the current table you have.
CLEANUP-CHANGES: Set the ReadFile() to read in a different way.  I have yet to actually have it in use, but it iterates through the properties and reads them out that way.

NOTE: It may still have problems, hence the branch creation.  I need to make adjustments so that it saves to the file correctly.

NOTE: There could potentially be an alorgithm to convert exp into level.  However, I did not find it quickly so I beared it no mind.  Worse comes to worse, make a HashMap or Dictionary and compare the values like that.